Move DustLocalStateError to errors file, implement Display trait
